Denormalize user information from ideas and improvements
Originally, ideas and improvements were associated with a
user account, and a single name and contact were associated
with all posts by that user. Based on requests to submit
ideas for other people, this no longer makes sense. So
columns for name and contact should be added to the idea
table and a contact column for improvements. For migration,
they should be filled with the name on the account.